#162857303 Validate all User Inputs

+++ app/api/v1/views/question_views.py
@@ -1,6 +1,7 @@
 """Question views"""
 from flask import make_response, jsonify, request, Blueprint
 from app.api.v1.models.question_models import QuestionModels
+from app.api.v1.utils.validation import verify_id
 
 questionv1 = Blueprint('questionv1', __name__, url_prefix='/api/v1')
 questions = QuestionModels()
@@ -9,14 +10,22 @@ questions = QuestionModels()
 @questionv1.route('/questions', methods=['POST'])
 def ask_question():
     """Create question"""
-    data = request.get_json()
-    user_id = data['user_id']
-    title = data['title']
-    body = data['body']
-    tags = data['tags']
-    answer = data['answer']
-
-    questions.add_question(user_id, title, body, tags, answer)
+    try:
+        data = request.get_json()
+        user_id = data['user_id']
+        title = data['title']
+        body = data['body']
+        tags = data['tags']
+        answer = data['answer']
+    except KeyError:
+        return make_response(jsonify({
+            "Error": "Please make sure all the fields are present"
+        }), 401)
+    qdata = questions.add_question(user_id, title, body, tags, answer)
+    if qdata == "empty":
+        return make_response(jsonify({
+                "Error": "All fields are required"
+            }), 401)
     return make_response(jsonify({
             "message": "Question Posted Successfully"
         }), 201)
@@ -34,6 +43,11 @@ def retrieve_questions():
 @questionv1.route('/questions/<questionId>', methods=['GET'])
 def retrieve_one_question(questionId):
     """Return one question"""
+    verify = verify_id(questionId)
+    if verify is False:
+            return make_response(jsonify({
+                "Error": "Please enter a valid question ID"
+            }), 400)
     one_question = questions.get_one_question(questionId)
     if one_question:
         return make_response(jsonify({
@@ -48,6 +62,11 @@ def retrieve_one_question(questionId):
 @questionv1.route('/questions/<questionId>', methods=['DELETE'])
 def delete_question(questionId):
     """Delete a question"""
+    verify = verify_id(questionId)
+    if verify is False:
+            return make_response(jsonify({
+                "Error": "Please enter a valid question ID"
+            }), 400)
     delete_question = questions.delete_one_question(questionId)
     if delete_question is False:
         return make_response(jsonify({
